input-shaping/requirements.txt
# Python Dependencies voor Input Shaping Analyse
# Werkplaats Marc - ADXL345 Resonantie Kalibratie
# GitHub: https://github.com/werkplaatsmarc/input-shaping-ender3

# Installatie: pip install -r requirements.txt

# Seriële communicatie met ADXL345
pyserial>=3.5

# Numerieke berekeningen
numpy>=1.21.0

# Grafiek visualisatie
matplotlib>=3.5.0

# Signaal verwerking en FFT analyse
scipy>=1.7.0